Provide header for all executioners
